Add testing for eventhandler and getnamehandler
