+++ package.json
@@ -4,6 +4,7 @@
   "description": "JS capstone project build with phaser",
   "main": "src/index.js",
   "scripts": {
+    "test": "jest",
     "build": "webpack --config webpack/prod.js ",
     "start": "webpack-dev-server --config webpack/base.js --open"
   },
@@ -19,6 +20,7 @@
   },
   "homepage": "https://github.com/ashraffares/Phaser_Game/#readme",
   "devDependencies": {
+    "jest": "^26.6.3",
     "@babel/core": "^7.7.2",
     "@babel/preset-env": "^7.7.1",
     "babel-loader": "^8.0.6",
