+++ src/localStorage/getNameHandler.js
@@ -1,0 +1,6 @@
+const getNameHandler = () => {
+  const name = localStorage.getItem('name');
+  return JSON.parse(name);
+};
+
+export default getNameHandler;
