+++ src/localStorage/saveHandler.js
@@ -1,0 +1,6 @@
+const saveHandler = (name) => {
+  localStorage.setItem('name', JSON.stringify(name));
+  return true;
+};
+
+export default saveHandler;
